[ScheduleDAG] Fix bug in check for use of dead defs
Was asserting when checking for uses of dead defs as it wasn't allowing for
subregisters
This was shown in the case (for AMDGPU) where:

%vreg79:sub3<def,dead,tied3> = V_MAC_F32_e32 %vreg914<undef>, %vreg901,
 %vreg79:sub3<tied0>, %EXEC<imp-use>;

and vreg79:sub0 vreg79:sub1 are subseqently used, but sub3 is not  - firing the
assertion incorrectly.

Modified the check to make sure that it only checks for uses of the same lane.

Subscribers: llvm-commits, MatzeB, tpr

Differential Revision: https://reviews.llvm.org/D34889