[RegisterCoalescer] Workaround for subrange with unresolved conflict
resolveConflicts() cannot cope with a subrange being marked
CR_Unresolved by analyzeValue(). I have a case where that happens, and I
think it is linked to the enclosing LR being marked as CR_Unresolved and
then being resolved by resolveConflicts. So I have disabled coalescing
in that case.

The reproducer is the best that bugpoint could do.

Change-Id: Iabf51a99255b44cc70a97e060474df435aef84e8